Update build.sh
The method of launching commands from PIP Install + Buildozer has been changed. Thanks to this, Buildozer operates within Virtalenv, not from the global system - which solves the problems with PEP 668. There is no more installation of the Buildozer via PIP in the container at the start, because everything happens in Virtalenv during the image build.